+++ ios/samples/KMSample1/KMSample1/AppDelegate.swift
@@ -15,8 +15,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
   func application(_ application: UIApplication,
                    didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]? = nil) -> Bool {
-    KeymanEngine.log.outputLevel = .debug
-    KeymanEngine.log.logAppDetails()
     return true
   }
 }