+++ mac/KeymanEngine4Mac/KeymanEngine4MacTests/KMEngineTests.m
@@ -41,21 +41,27 @@ NSString * names[nCombinations];
     KMXFile *kmxFile = [KeymanEngineTestsStaticHelperMethods getKmxFileTestMacEngine];
     KMEngine *engine = [[KMEngine alloc] initWithKMX:kmxFile context:@"" verboseLogging:YES];
     XCTAssert(engine != nil, @"Expected non-nil engine");
-    XCTAssert(engine.getCoreContext.length == 0, @"Expected empty context buffer");
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([engine.getCoreContextDebug isEqualToString:@"|| (len: 0) [ ]"], @"Expected empty context buffer");
 }
 
 - (void)testinitWithKMX_ValidKmxNonEmptyContext_InitializedWithContext {
     KMXFile *kmxFile = [KeymanEngineTestsStaticHelperMethods getKmxFileTestMacEngine];
     KMEngine *engine = [[KMEngine alloc] initWithKMX:kmxFile context:@"abc" verboseLogging:YES];
     XCTAssert(engine != nil, @"Expected non-nil engine");
-    XCTAssert([engine.getCoreContext isEqualToString:@"abc"], @"Expected 'abc' in context buffer");
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([engine.getCoreContextDebug isEqualToString:@"|abc| (len: 3) [ U+0061 U+0062 U+0063 ]"], @"Expected 'abc' in context buffer");
 }
 
 - (void)testsetCoreContextIfNeeded_NonEmptyContext_InitialContextUpdated {
     KMXFile *kmxFile = [KeymanEngineTestsStaticHelperMethods getKmxFileTestMacEngine];
     KMEngine *engine = [[KMEngine alloc] initWithKMX:kmxFile context:@"a" verboseLogging:YES];
     [engine setCoreContextIfNeeded:@"xyz"];
-    XCTAssert([engine.getCoreContext isEqualToString:@"xyz"], @"Expected 'xyz' in context buffer");
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([engine.getCoreContextDebug isEqualToString:@"|xyz| (len: 3) [ U+0078 U+0079 U+007a ]"], @"Expected 'xyz' in context buffer");
 }
 
 // TODO: re-enable this one after the core API km_core_state_context_set_if_needed is fixed
@@ -64,7 +70,9 @@ NSString * names[nCombinations];
     KMXFile *kmxFile = [KeymanEngineTestsStaticHelperMethods getKmxFileTestMacEngine];
     KMEngine *engine = [[KMEngine alloc] initWithKMX:kmxFile context:@"" verboseLogging:YES];
     [engine setCoreContextIfNeeded:@"xyz"];
-    XCTAssert([engine.getCoreContext isEqualToString:@"xyz"], @"Expected 'xyz' in context buffer");
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([engine.getCoreContextDebug isEqualToString:@"|xyz| (len: 3) [ U+0078 U+0079 U+007a ]"], @"Expected 'xyz' in context buffer");
 }
  */
 
@@ -278,7 +286,7 @@ NSString * names[nCombinations];
         NSString * characters = charactersIgnoringModifiers;
         if (modifiers[i] & (LEFT_ALT_FLAG | RIGHT_ALT_FLAG))
             characters = [characters stringByAppendingString:@"\u030A"];
-        
+
         NSLog(@"Test case: %lu", (NSUInteger)modifiers[i]);
         // NOTE: 'a' happens to be keyCode 0 (see initVirtualKeyMapping in CoreHelper)
         NSEvent *event = [NSEvent keyEventWithType:NSEventTypeKeyDown location:NSMakePoint(0, 0) modifierFlags:modifiers[i] timestamp:0 windowNumber:0 context:nil characters:characters charactersIgnoringModifiers:charactersIgnoringModifiers isARepeat:NO keyCode:0];
@@ -335,10 +343,10 @@ NSString * names[nCombinations];
 - (void)testprocessEvent_eventForSWithModifiers_ReturnsCharacterActionWithExpectedCharacterBasedOnKmx {
     int i = 0;
     [KMEngineTests fillInNamesAndModifiersForAllChiralCombinations];
-    
+
     KMXFile *kmxFile = [KeymanEngineTestsStaticHelperMethods getKmxFileTestMacEngine];
     KMEngine *engine = [[KMEngine alloc] initWithKMX:kmxFile context:@"" verboseLogging:YES];
-    
+
     for (i = 0; i < nCombinations; i++) {
         [engine clearCoreContext];
         NSString *charactersIgnoringModifiers = (modifiers[i] & (LEFT_SHIFT_FLAG | RIGHT_SHIFT_FLAG)) ? @"S" : @"s";
@@ -349,7 +357,7 @@ NSString * names[nCombinations];
             else
                 characters = @"ß";
         }
-        
+
         NSLog(@"Test case: %lu", (NSUInteger)modifiers[i]);
         // NOTE: 's' happens to be keyCode 1 (see initVirtualKeyMapping in CoreHelper)
         NSEvent *event = [NSEvent keyEventWithType:NSEventTypeKeyDown location:NSMakePoint(0, 0) modifierFlags:modifiers[i] timestamp:0 windowNumber:0 context:nil characters:characters charactersIgnoringModifiers:charactersIgnoringModifiers isARepeat:NO keyCode:1];
@@ -554,8 +562,10 @@ NSString * names[nCombinations];
     NSLog(@"output: %@", output);
     XCTAssert(output.codePointsToDeleteBeforeInsert == 1, @"Expected output to delete one code point");
     XCTAssert(!output.hasTextToInsert, @"expected to insert nothing");
-    NSString *context = engine.getCoreContext;
-    XCTAssert([context isEqualToString:@""], @"Context should be empty.");
+    NSString *context = engine.getCoreContextDebug;
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([context isEqualToString:@"|| (len: 0) [ ]"], @"Context should be empty.");
 }
 
 - (void)testCoreProcessEvent_eventReturnWithElNuerKmx_EmitWithContextEmpty {
@@ -564,8 +574,10 @@ NSString * names[nCombinations];
     NSEvent *event = [NSEvent keyEventWithType:NSEventTypeKeyDown location:NSMakePoint(0, 0) modifierFlags:0 timestamp:0 windowNumber:0 context:nil characters:@"\n" charactersIgnoringModifiers:@"\n" isARepeat:NO keyCode:kVK_Return];
     CoreKeyOutput *output = [engine processEvent:event];
     XCTAssert(output.emitKeystroke, @"Expected emitKeystroke==YES");
-    NSString *context = engine.getCoreContext;
-    XCTAssert([context isEqualToString:@""], @"Context should be cleared.");
+    NSString *context = engine.getCoreContextDebug;
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([context isEqualToString:@"|| (len: 0) [ ]"], @"Context should be cleared.");
 }
 
 - (void)testCoreProcessEvent_eventTabWithElNuerKmx_EmitWithContextEmpty {
@@ -574,8 +586,10 @@ NSString * names[nCombinations];
     NSEvent *event = [NSEvent keyEventWithType:NSEventTypeKeyDown location:NSMakePoint(0, 0) modifierFlags:0 timestamp:0 windowNumber:0 context:nil characters:@"\t" charactersIgnoringModifiers:@"\t" isARepeat:NO keyCode:kVK_Tab];
     CoreKeyOutput *output = [engine processEvent:event];
     XCTAssert(output.emitKeystroke, @"Expected emitKeystroke==YES");
-    NSString *context = engine.getCoreContext;
-    XCTAssert([context isEqualToString:@""], @"Context should be cleared.");
+    NSString *context = engine.getCoreContextDebug;
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([context isEqualToString:@"|| (len: 0) [ ]"], @"Context should be cleared.");
 }
 
 - (void)testCoreProcessEvent_eventSingleQuoteWithElNuerKmx_ReturnsDiacritic {
@@ -585,8 +599,10 @@ NSString * names[nCombinations];
     NSEvent *event = [NSEvent keyEventWithType:NSEventTypeKeyDown location:NSMakePoint(0, 0) modifierFlags:0 timestamp:0 windowNumber:0 context:nil characters:@"'" charactersIgnoringModifiers:@"'" isARepeat:NO keyCode:kVK_ANSI_Quote];
     CoreKeyOutput *output = [engine processEvent:event];
     XCTAssert(output.hasTextToInsert, @"returns text to insert");
-    context = engine.getCoreContext;
-    XCTAssert([context isEqualToString:@"\u025B\u0308"], @"Context updated with diacritic.");
+    context = engine.getCoreContextDebug;
+    // Note: relying on km_core_state_context_debug output format is just barely
+    // acceptable for a unit test
+    XCTAssert([context isEqualToString:@"|\u025B\u0308| (len: 2) [ U+025b U+0308 ]"], @"Context updated with diacritic.");
 }
 
 @end