+++ web/src/engine/main/src/hardKeyboard.ts
@@ -39,7 +39,9 @@ export function processForMnemonicsAndLegacy(s: KeyEvent, activeKeyboard: Keyboa
     //
     // Third:  DO, however, track direct presses of any main modifier key.  The OSK should
     // reflect the current modifier state even for legacy keyboards.
-    if(!activeKeyboard.definesPositionalOrMnemonic && !(s.Lmodifiers & 0x60) && !s.isModifier) {
+    if(!activeKeyboard.definesPositionalOrMnemonic &&
+       !(s.Lmodifiers & Codes.modifierBitmasks.NON_LEGACY) &&
+       !s.isModifier) {
       // Support version 1.0 KeymanWeb keyboards that do not define positional vs mnemonic
       s = new KeyEvent({
         Lcode: KeyMapping._USKeyCodeToCharCode(s),